Create README.md
On March 2020, WHO has declared COVID-19 as a pandemic. Since then many countries have declared complete lockdown. Do you know how people are feeling and what they are going through in this isolation period? This period is not only a threat to people's physical health but due to the social distancing and a cut from the real-world, mental health is being endangered too.
Using the trending tweets related to COVID- 19, we will build a sentiment analyzer to know how people are perceiving the worldwide lockdown. Our model can aid in post-crisis management of human resources for a better and happy future ahead.